move insert count / cache counts
